pyproject.toml
[project]
name = "fact-checking-news-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.11",
    "groq>=0.29.0",
    "jenkins>=1.0.2",
    "python-a2a>=0.5.9",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
